requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
Django==3.2
entrypoints==0.4
filelock==3.9.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
pipenv==2023.2.18
platformdirs==3.0.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
streamlit==1.21.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
